// Frontend/src/app/modules/features/code-tables/enums/error-message.enum.ts
export enum ErrorMessageEnum {
  GET = 'Unable to get data at this time and no further information is available.',
  PUT = 'Unable to get update the record at this time and no further information is available.',
  POST = 'Unable to insert the new record at this time and no further information is available.',
  DELETE = 'Unable to delete the record at this time and no further information is available.',
}