// Frontend/src/app/modules/shared/enums/form-field-type.enum.ts
export enum FormFieldTypeEnum {
  Input = 'input',
  TextArea = 'textarea',
  Select = 'select',
  Autocomplete = 'autocomplete',
  Date = 'date',
  DateTime = 'datetime',
  MonthDayDate = 'month_day_date',
  DateOrDateTime = 'datetimeoptional',
  Currency = 'currency',
  Checkbox = 'checkbox',
  File = 'file',
}